refactor(views): 共享连接状态条 helper 抽到 views/connection-status-bar.js
Phase 3 第一步。把 Phase 2 加入的三个 helper（formatWechatsyncCheckedAt /
describeWechatsyncConnectionState / renderWechatsyncConnectionStatusBar）
从 input.js 移出，建立 views/ 目录承载 UI 渲染逻辑。input.js 改为 require
新模块、复用同名 export，行为不变。

后续 Phase 3 会陆续把「其他平台」settings tab、发布弹窗等大块也搬过来，
最终目标是把 input.js 拆成核心逻辑 + views/ 视图层。

测试无变化，51 / 594 全过。
